Support non-standard named kwargs.
This means checking for the ** rather than the "kwargs", which is more
robust and correct in other ways too.

Unfortunately, this would mean adding a "last_last_last_token" and thats
starting to get nasty, so this refactors the checks to just inspect the
list of tokens and jump positions in that.

Comments have to be filtered out beforehand otherwise they could be an
arbitrary number of tokens between `idx` and the ** or newline we're
checking for.