Support tokens within f-strings in Python 3.12
In Python 3.12 token streams provide details of tokens within
f-strings, rather than merely presenting these as strings. This
means that simple string comparisons of tokens are now somewhat
dangerous since a token with string ')' can no longer be sure to
be a closing paren which is part of syntax -- it may now in fact
be part of the text of an f-string which happens to be only that
character.